# Source/Zybach.API/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Zybach",
    "version": "v1"
  },
  "paths": {
...
    "/api/chemigationPermits/{chemigationPermitNumber}/{recordYear}/chemicalFormulations": {
      "get": {
        "tags": [
          "ChemigationPermitAnnualRecordChemicalFormulation"
        ],
        "parameters": [
          {
            "name": "chemigationPermitNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "recordYear",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChemigationPermitAnnualRecordChemicalFormulationSimpleDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChemigationPermitAnnualRecordChemicalFormulationSimpleDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChemigationPermitAnnualRecordChemicalFormulationSimpleDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChemigationPermitAnnualRecordChemicalFormulationSimpleDto": {
        "type": "object",
        "properties": {
          "ChemicalFormulationName": {
            "type": "string",
            "nullable": true
          },
          "ChemicalUnitName": {
            "type": "string",
            "nullable": true
          },
          "ChemigationPermitAnnualRecordChemicalFormulationID": {
            "type": "integer",
            "format": "int32"
          },
          "ChemigationPermitAnnualRecordID": {
            "type": "integer",
            "format": "int32"
          },
          "ChemicalFormulationID": {
            "type": "integer",
            "format": "int32"
          },
          "ChemicalUnitID": {
            "type": "integer",
            "format": "int32"
          },
          "TotalApplied": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "AcresTreated": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}